requirements.txt
requests==2.28.2
python-dotenv==1.0.0
python-telegram-bot==13.0

